+++ app/domain/document.model.ts
@@ -2,6 +2,7 @@ import * as z from 'zod';
 
 export enum DocumentType {
   REGISTRATION_CERTIFICATE = 'registrationCertificate',
+  INSPECTION_CERTIFICATE = 'inspectionCertificate',
   OTHER = 'other',
 }